Add files via upload
This is Pharmacy Management System in which we used HTML and CSS as a Front End Language and PHP as a back-end language. The login page shows Admin, Manager, Pharmacist, and Cashier portals. Each of them have a separate login and password. Next PHP files show different features of Admin, Pharmacist, Cashier and Manager.
Database is in Pharmacy.sql file.